refactor: Extract vue node pointer events into a composable.
